Automatically reformat all files, apply code cleanup

+++ src/main/java/com/github/unaimillan/rars/assembler/Tokenizer.java
@@ -201,21 +201,21 @@ public class Tokenizer {
      * @return the generated token list for that line
      **/
     /*
-    * 
-    * Tokenizing is not as easy as it appears at first blush, because the typical 
-    * delimiters: space, tab, comma, can all appear inside quoted ASCII strings!
-    * Also, spaces are not as necessary as they seem, the following line is accepted
-    * and parsed correctly by SPIM:    label:lw,$t4,simple#comment
-    * as is this weird variation:      label  :lw  $t4  ,simple ,  ,  , # comment
-    * 
-    * as is this line:  stuff:.asciiz"# ,\n\"","aaaaa"  (interestingly, if you put
-    * additional characters after the \", they are ignored!!)
-    * 
-    * I also would like to know the starting character position in the line of each
-    * token, for error reporting purposes.  StringTokenizer cannot give you this.
-    * 
-    * Given all the above, it is just as easy to "roll my own" as to use StringTokenizer
-    */
+     *
+     * Tokenizing is not as easy as it appears at first blush, because the typical
+     * delimiters: space, tab, comma, can all appear inside quoted ASCII strings!
+     * Also, spaces are not as necessary as they seem, the following line is accepted
+     * and parsed correctly by SPIM:    label:lw,$t4,simple#comment
+     * as is this weird variation:      label  :lw  $t4  ,simple ,  ,  , # comment
+     *
+     * as is this line:  stuff:.asciiz"# ,\n\"","aaaaa"  (interestingly, if you put
+     * additional characters after the \", they are ignored!!)
+     *
+     * I also would like to know the starting character position in the line of each
+     * token, for error reporting purposes.  StringTokenizer cannot give you this.
+     *
+     * Given all the above, it is just as easy to "roll my own" as to use StringTokenizer
+     */
 
     // Modified for release 4.3, to preserve existing API.
     public TokenList tokenizeLine(int lineNum, String theLine) {
@@ -436,7 +436,7 @@ public class Tokenizer {
             linePos++;
         }  // while
         if (tokenPos > 0) {
-            if(insideQuotedString){
+            if (insideQuotedString) {
                 errors.add(new ErrorMessage(program, lineNum, tokenStartPos,
                         "String is not terminated."));
             }
@@ -557,7 +557,7 @@ public class Tokenizer {
         String quotesRemoved = value.substring(1, value.length() - 1);
         // if not escaped, then if one character left return its value else return original.
         if (quotesRemoved.charAt(0) != '\\') {
-            return (quotesRemoved.length() == 1) ? Integer.toString((int) quotesRemoved.charAt(0)) : value;
+            return (quotesRemoved.length() == 1) ? Integer.toString(quotesRemoved.charAt(0)) : value;
         }
         // now we know it is escape sequence and have to decode which of the 8: ',",\,n,t,b,r,f
         if (quotesRemoved.length() == 2) {
