+++ src/main/java/com/github/unaimillan/rars/riscv/instructions/ImmediateInstruction.java
@@ -1,10 +1,10 @@
 package com.github.unaimillan.rars.riscv.instructions;
 
 import com.github.unaimillan.rars.ProgramStatement;
-import com.github.unaimillan.rars.riscv.InstructionSet;
-import com.github.unaimillan.rars.riscv.hardware.RegisterFile;
 import com.github.unaimillan.rars.riscv.BasicInstruction;
 import com.github.unaimillan.rars.riscv.BasicInstructionFormat;
+import com.github.unaimillan.rars.riscv.InstructionSet;
+import com.github.unaimillan.rars.riscv.hardware.RegisterFile;
 
 /*
 Copyright (c) 2017,  Benjamin Landers
@@ -44,17 +44,18 @@ public abstract class ImmediateInstruction extends BasicInstruction {
         super(usage, description, BasicInstructionFormat.I_FORMAT,
                 "tttttttttttt sssss " + funct + " fffff 0010011");
     }
+
     public ImmediateInstruction(String usage, String description, String funct, boolean rv64) {
         super(usage, description, BasicInstructionFormat.I_FORMAT,
-                "tttttttttttt sssss " + funct + " fffff 0011011",rv64);
+                "tttttttttttt sssss " + funct + " fffff 0011011", rv64);
     }
 
     public void simulate(ProgramStatement statement) {
         int[] operands = statement.getOperands();
-        if (InstructionSet.rv64){
+        if (InstructionSet.rv64) {
             RegisterFile.updateRegister(operands[0], compute(RegisterFile.getValueLong(operands[1]),
-                    (operands[2] << 20) >> 20)); // make sure the immediate is sign-extended
-        }else {
+                    ((long) operands[2] << 20) >> 20)); // make sure the immediate is sign-extended
+        } else {
             RegisterFile.updateRegister(operands[0], computeW(RegisterFile.getValue(operands[1]),
                     (operands[2] << 20) >> 20)); // make sure the immediate is sign-extended
         }
@@ -66,12 +67,13 @@ public abstract class ImmediateInstruction extends BasicInstruction {
      * @return the result to be stored from the instruction
      */
     protected abstract long compute(long value, long immediate);
+
     /**
      * @param value     the truncated value from the register
      * @param immediate the value from the immediate
      * @return the result to be stored from the instruction
      */
-    protected int computeW(int value, int immediate){
-        return (int) compute(value,immediate);
+    protected int computeW(int value, int immediate) {
+        return (int) compute(value, immediate);
     }
 }
